refactorizado

+++ src/main/java/com/comedor/dao/IRolDAO.java
@@ -1,12 +1,8 @@
 package com.comedor.dao;
 
 import org.springframework.data.jpa.repository.JpaRepository;
+import com.comedor.dto.Rol;
 
-import comedor.dto.Rol;
-
-
-
-public interface IRolDAO extends JpaRepository<Rol, Integer>{
-	
+public interface IRolDAO extends JpaRepository<Rol, Long>{
 		
 }
