+++ src/main/java/com/comedor/dao/IUsuarioDAO.java
@@ -2,9 +2,8 @@ package com.comedor.dao;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 
-import comedor.dto.Usuario;
+import com.comedor.dto.Usuario;
 
-
-public interface IUsuarioDAO extends JpaRepository<Usuario, Integer>{
+public interface IUsuarioDAO extends JpaRepository<Usuario, Long>{
 	
 }
